Bug fix: fix cert decompression via zlib by ignoring EOF (#25)
An io.Reader is allowed (but not required) to return io.EOF upon reading the
last byte of the stream, even if no subsequent bytes were requested. We later
check that we read the expected number of bytes, so we can safely ignore EOF
errors returned by the decompression readers.